requirements.txt
acres==0.5.0
aiofiles==24.1.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.0.1
cloudinary==1.44.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
dnspython==2.7.0
email-validator==2.0.0
etelemetry==0.3.1
exceptiongroup==1.3.0
filelock==3.18.0
fitz==0.0.1.dev2
Flask==2.0.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
frontend==0.0.3
gunicorn==20.1.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.0.1
Jinja2==3.0.1
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==2.0.1
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
numpy==2.0.2
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
Pillow==10.0.0
prov==2.1.1
puremagic==1.29
pydot==4.0.1
pymongo==4.0.1
pyparsing==3.2.3
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2025.2
pyxnat==1.6.3
svglib==1.5.1
cairosvg==2.7.1
qrcode==7.4.2
rdflib==7.1.4
reportlab==4.0.4
requests==2.32.4
scipy==1.13.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
tools==1.0.2
traits==7.0.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==2.0.1
WTForms==3.2.1
zipp==3.23.0